+++ HidingBar.lua
@@ -13,11 +13,16 @@ hidingBar.drag:SetHitRectInsets(-2, -2, -2, -2)
 hidingBar.drag:SetFrameLevel(hidingBar:GetFrameLevel())
 hidingBar.drag.bg = hidingBar.drag:CreateTexture(nil, "OVERLAY")
 hidingBar.drag.bg:SetAllPoints()
+hidingBar.drag.fTimer = CreateFrame("FRAME")
 hidingBar.createdButtons, hidingBar.minimapButtons, hidingBar.mixedButtons = {}, {}, {}
 local createdButtonsByName, btnSettings = {}, {}
+local ldb = LibStub("LibDataBroker-1.1")
+local ldbi, ldbi_ver = LibStub("LibDBIcon-1.0")
+local MSQ = LibStub("Masque", true)
 
 
 local ignoreFrameList = {
+	["LibDBIcon10_HidingBar"] = true,
 	["GameTimeFrame"] = true,
 	["QueueStatusMinimapButton"] = true,
 	["HelpOpenTicketButton"] = true,
@@ -101,8 +106,19 @@ function hidingBar:ADDON_LOADED(addonName)
 		self.config.ignoreMBtn = self.config.ignoreMBtn or {"GatherMatePin"}
 		self.config.bgColor = self.config.bgColor or {.1, .1, .1, .7}
 		self.config.lineColor = self.config.lineColor or {.8, .6, 0}
+		self.config.omb = self.config.omb or {
+			hide = true,
+			anchor = "right",
+			size = 31,
+			lock = self.config.lock,
+		}
 		self.config.btnSettings = setmetatable(self.config.btnSettings or {}, meta)
 		self.config.mbtnSettings = setmetatable(self.config.mbtnSettings or {}, meta)
+
+		if self.config.freeMove then
+			self.config.barTypePosition = 1
+			self.config.freeMove = nil
+		end
 
 		self:setFrameStrata()
 		self.bg:SetVertexColor(unpack(self.config.bgColor))
@@ -117,10 +133,47 @@ function hidingBar:ADDON_LOADED(addonName)
 end
 
 
+function hidingBar:createOwnMinimapButton()
+	self.ldb_icon = ldb:NewDataObject(addon, {
+		type = "data source",
+		text = addon,
+		icon = "Interface/Icons/misc_arrowleft",
+		OnClick = function(_, button)
+			if button == "LeftButton" then
+				if self:IsShown() then
+					self:Hide()
+				else
+					local func = self.drag:GetScript("OnClick")
+					if func then func(self.drag) end
+				end
+			elseif button == "RightButton" then
+				if IsAltKeyDown() then
+					self.config.lock = not self.config.lock
+					self:refreshShown()
+					if config.lock then config.lock:SetChecked(self.config.lock) end
+				end
+				if IsShiftKeyDown() then
+					config:openConfig()
+				end
+			end
+		end,
+		OnEnter = function()
+			local func = self.drag:GetScript("OnEnter")
+			if func then func(self.drag) end
+		end,
+		OnLeave = function(btn)
+			local func = self.drag:GetScript("OnLeave")
+			if func then func(self.drag) end
+		end,
+	})
+	ldbi:Register(addon, self.ldb_icon, self.config.omb)
+end
+
+
 function hidingBar:UI_SCALE_CHANGED()
 	self.position = nil
 	self.secondPosition = nil
-	self:setBarPosition()
+	self:updateBarPosition()
 end
 
 
@@ -136,7 +189,6 @@ end
 function hidingBar:init()
 	local t = time()
 
-	local MSQ = LibStub("Masque", true)
 	if MSQ then
 		self.MSQ_Button = MSQ:Group(addon, L["DataBroker Buttons"], "DataBroker")
 		self.MSQ_Button:SetCallback(function()
@@ -187,7 +239,6 @@ function hidingBar:init()
 	end
 
 	if self.config.grabMinimap then
-		local ldbi, ldbi_ver = LibStub("LibDBIcon-1.0", true)
 		if ldbi and ldbi_ver >= 39 then
 			local ldbiTbl = ldbi:GetButtonList()
 			for i = 1, #ldbiTbl do
@@ -580,12 +631,14 @@ function hidingBar:init()
 	end
 	self.db.tstmp = t
 
+	self:createOwnMinimapButton()
+
 	self.drag:setShowHandler()
 	self:sort()
 	self:setButtonSize()
 	self:applyLayout()
-	self:setBarPosition()
-	self:setDragBarPosition()
+	self:setBarTypePosition(self.config.barTypePosition)
+	self:updateDragBarPosition()
 
 	self:RegisterEvent("UI_SCALE_CHANGED")
 end
@@ -784,7 +837,7 @@ function hidingBar:grabMinimapAddonsButtons(parentFrame, t)
 end
 
 
-function hidingBar:setMButtonRegions(btn)
+function hidingBar:setMButtonRegions(btn, getData)
 	local name, texture, layer, border, background, icon, highlight, data
 	for _, region in ipairs({btn:GetRegions()}) do
 		if region:GetObjectType() == "Texture" then
@@ -816,6 +869,8 @@ function hidingBar:setMButtonRegions(btn)
 	elseif highlight then
 		data = {Highlight = highlight}
 	end
+
+	if getData then return data end
 	self.MSQ_MButton:AddButton(btn, data, nil, true)
 	self:MSQ_MButton_Update(btn)
 end
@@ -980,6 +1035,17 @@ function hidingBar:applyLayout()
 end
 
 
+function hidingBar:setLocked(lock)
+	self.config.lock = lock
+	self:refreshShown()
+	if lock then
+		ldbi:Lock(addon)
+	else
+		ldbi:Unlock(addon)
+	end
+end
+
+
 function hidingBar:setFrameStrata()
 	local strata
 	if self.config.frameStrata == 5 then
@@ -1001,7 +1067,7 @@ function hidingBar:setFrameStrata()
 end
 
 
-function hidingBar:setDragBarPosition()
+function hidingBar:updateDragBarPosition()
 	local anchor = self.config.anchor
 	self.drag:ClearAllPoints()
 	if self:IsShown() then
@@ -1036,8 +1102,15 @@ function hidingBar:setDragBarPosition()
 end
 
 
+function hidingBar:setMBAnchor(anchor)
+	if self.config.barTypePosition ~= 2 or self.config.omb.anchor == anchor then return end
+	self.config.omb.anchor = anchor
+	self:setBarTypePosition(self.config.barTypePosition)
+end
+
+
 function hidingBar:setBarAnchor(anchor)
-	if not self.config.freeMove or self.config.anchor == anchor then return end
+	if self.config.barTypePosition ~= 1 or self.config.anchor == anchor then return end
 	local x, y, position, secondPosition = self:GetCenter()
 	self.config.anchor = anchor
 	local width, height = self:applyLayout()
@@ -1071,13 +1144,15 @@ function hidingBar:setBarAnchor(anchor)
 		secondPosition = y - height
 	end
 
-	self:setBarPosition(position, secondPosition)
+	self:setBarCoords(position, secondPosition)
+	self:updateBarPosition()
 end
 
 
 function hidingBar:setBarExpand(expand)
 	if self.config.expand == expand then return end
 	local anchor, delta, position = self.config.anchor
+	local scale = UIParent:GetScale()
 
 	if anchor == "left" or anchor == "right" then
 		delta = self:GetHeight()
@@ -1091,13 +1166,93 @@ function hidingBar:setBarExpand(expand)
 	end
 
 	if expand == 0 then
-		position = self.position + delta
+		position = self.config.position / scale + delta
 	else
-		position = self.position - delta
+		position = self.config.position / scale - delta
 	end
 	self.config.expand = expand
 
-	self:setBarPosition(position)
+	self:setBarCoords(position)
+	self:setBarTypePosition(self.config.barTypePosition)
+end
+
+
+function hidingBar:setBarTypePosition(typePosition)
+	self.config.barTypePosition = typePosition
+
+	if self.config.barTypePosition == 2 then
+		ldbi:Show(addon)
+		if not self.mb then
+			self.mb = ldbi:GetMinimapButton(addon)
+			if MSQ then
+				self.MSQ_OMB = MSQ:Group(addon, L["Own Minimap Button"], "OMB")
+				self.MSQ_OMB:SetCallback(function() self:MSQ_MButton_Update(self.mb) end)
+				self.MSQ_OMB:AddButton(self.mb, self:setMButtonRegions(self.mb, true), nil, true)
+				self:MSQ_MButton_Update(self.mb)
+			end
+		end
+		local btnSize, position, secondPosition = 31
+
+		if self.config.omb.anchor == "left" or self.config.omb.anchor == "right" then
+			if self.config.expand == 0 then
+				position = btnSize
+			elseif self.config.expand == 1 then
+				position = 0
+			else
+				position = btnSize / 2
+			end
+		else
+			if self.config.expand == 0 then
+				position = 0
+			elseif self.config.expand == 1 then
+				position = btnSize
+			else
+				position = btnSize / 2
+			end
+		end
+
+		if self.config.omb.anchor == "left" then
+			secondPosition = btnSize
+			self.ldb_icon.icon = "Interface/Icons/misc_arrowright"
+		elseif self.config.omb.anchor == "right" then
+			secondPosition = -btnSize
+			self.ldb_icon.icon = "Interface/Icons/misc_arrowleft"
+		elseif self.config.omb.anchor == "top" then
+			secondPosition = -btnSize
+			self.ldb_icon.icon = "Interface/Icons/misc_arrowdown"
+		else
+			secondPosition = btnSize
+			self.ldb_icon.icon = "Interface/Icons/misc_arrowlup"
+		end
+
+		self.anchorObj = self.config.omb
+		self.rFrame = self.mb
+		self.position = position
+		self.secondPosition = secondPosition
+	else
+		ldbi:Hide(addon)
+		self.anchorObj = self.config
+		self.rFrame = UIParent
+		self.position = nil
+		self.secondPosition = nil
+	end
+
+	self:updateBarPosition()
+end
+
+
+function hidingBar:setBarCoords(position, secondPosition)
+	local scale = UIParent:GetScale()
+
+	if position then
+		self.position = position
+		self.config.position = position * scale
+	end
+
+	if secondPosition then
+		self.secondPosition = secondPosition
+		self.config.secondPosition = secondPosition * scale
+	end
 end
 
 
@@ -1109,20 +1264,9 @@ do
 		bottom = {[0] = "BOTTOMLEFT", "BOTTOMRIGHT", "BOTTOM"},
 	}
 
-	function hidingBar:setBarPosition(position, secondPosition)
+	function hidingBar:updateBarPosition()
 		local width, height = self:GetSize()
-		local scale = UIParent:GetScale()
-		local anchor = self.config.anchor
-
-		if position then
-			self.position = position
-			self.config.position = position * scale
-		end
-
-		if secondPosition then
-			self.secondPosition = secondPosition
-			self.config.secondPosition = secondPosition * scale
-		end
+		local anchor = self.anchorObj.anchor
 
 		if not self.position then
 			if not self.config.position then
@@ -1132,14 +1276,14 @@ do
 					self.config.position = WorldFrame:GetWidth() / 2
 				end
 			end
-			self.position = self.config.position / scale
+			self.position = self.config.position / UIParent:GetScale()
 		end
 
 		if not self.secondPosition then
 			if not self.config.secondPosition then
 				self.config.secondPosition = 0
 			end
-			self.secondPosition = self.config.secondPosition / scale
+			self.secondPosition = self.config.secondPosition / UIParent:GetScale()
 		end
 
 		config:updateCoords()
@@ -1147,13 +1291,13 @@ do
 		local point = pointForExpand[anchor][self.config.expand]
 		self:ClearAllPoints()
 		if anchor == "left" then
-			self:SetPoint(point, UIParent, "BOTTOMLEFT", self.secondPosition, self.position)
+			self:SetPoint(point, self.rFrame, "BOTTOMLEFT", self.secondPosition, self.position)
 		elseif anchor == "right" then
-			self:SetPoint(point, UIParent, "BOTTOMRIGHT", self.secondPosition, self.position)
+			self:SetPoint(point, self.rFrame, "BOTTOMRIGHT", self.secondPosition, self.position)
 		elseif anchor == "top" then
-			self:SetPoint(point, UIParent, "TOPLEFT", self.position, self.secondPosition)
+			self:SetPoint(point, self.rFrame, "TOPLEFT", self.position, self.secondPosition)
 		else
-			self:SetPoint(point, UIParent, "BOTTOMLEFT", self.position, self.secondPosition)
+			self:SetPoint(point, self.rFrame, "BOTTOMLEFT", self.position, self.secondPosition)
 		end
 	end
 end
@@ -1168,7 +1312,7 @@ function hidingBar:dragBar()
 	local scale = UIParent:GetScale()
 	x, y = x / scale, y / scale
 
-	if not self.config.freeMove then
+	if self.config.barTypePosition == nil then
 		local offset = 100
 
 		if not IsShiftKeyDown() then
@@ -1205,7 +1349,7 @@ function hidingBar:dragBar()
 		if anchor ~= self.config.anchor then
 			self.config.anchor = anchor
 			width, height = self:applyLayout()
-			self:setDragBarPosition()
+			self:updateDragBarPosition()
 
 			if config.hideToCombobox then
 				UIDropDownMenu_SetSelectedValue(config.hideToCombobox, self.config.anchor)
@@ -1223,7 +1367,7 @@ function hidingBar:dragBar()
 			delta = 0
 		end
 		position = y - delta
-		if self.config.freeMove then
+		if self.config.barTypePosition == 1 then
 			local dhWidth = self.drag:GetWidth() / 2
 			delta = anchor == "left" and width + dhWidth or UIwidth - width - dhWidth
 			secondPosition = x - delta
@@ -1238,14 +1382,15 @@ function hidingBar:dragBar()
 			delta = 0
 		end
 		position = x - delta
-		if self.config.freeMove then
+		if self.config.barTypePosition == 1 then
 			local dhHeight = self.drag:GetHeight() / 2
 			delta = anchor == "top" and UIheight - height - dhHeight or height + dhHeight
 			secondPosition = y - delta
 		end
 	end
 
-	self:setBarPosition(position, secondPosition)
+	self:setBarCoords(position, secondPosition)
+	self:updateBarPosition()
 end
 
 
@@ -1285,7 +1430,7 @@ function hidingBar:enter(force)
 		self:SetScript("OnUpdate", nil)
 		self:Show()
 		self:Raise()
-		self:setDragBarPosition()
+		self:updateDragBarPosition()
 	end
 end
 hidingBar:SetScript("OnEnter", hidingBar.enter)
@@ -1295,9 +1440,9 @@ function hidingBar:hideBar(elapsed)
 	self.timer = self.timer - elapsed
 	if self.timer <= 0 then
 		self:Hide()
-		self:setDragBarPosition()
+		self:updateDragBarPosition()
 		self:SetScript("OnUpdate", nil)
-		if self.config.fade then
+		if self.config.fade and self.drag:IsShown() then
 			UIFrameFadeOut(self.drag, 1.5, self.drag:GetAlpha(), self.config.fadeOpacity)
 		end
 	end
@@ -1315,24 +1460,33 @@ hidingBar:SetScript("OnLeave", hidingBar.leave)
 
 
 function hidingBar:refreshShown()
-	if self.config.showHandler ~= 3 then
+	if self.config.barTypePosition == 2 then
+		self.drag:Hide()
+		if self.config.showHandler == 3 then
+			self:enter(true)
+		end
+		if self:IsShown() then
+			self:leave()
+		end
+	elseif self.config.showHandler == 3 then
+		self:enter(true)
+		self.drag:SetShown(not self.config.lock and self.shown)
+	else
 		if self:IsShown() then
 			self:leave()
 		else
-			self:setDragBarPosition()
+			self:updateDragBarPosition()
 			if self.config.fade then
 				UIFrameFadeOut(self.drag, 1.5, self.drag:GetAlpha(), self.config.fadeOpacity)
 			end
 		end
 		self.drag:SetShown(self.shown)
-	else
-		self:enter(true)
-		self.drag:SetShown(not self.config.lock and self.shown)
 	end
 end
 
 
 function hidingBar.drag:hoverHandler()
+	if not hidingBar.drag:IsShown() then return end
 	if hidingBar.config.fade then
 		UIFrameFadeOut(self, hidingBar.config.showDelay, self:GetAlpha(), 1)
 	end
@@ -1352,9 +1506,9 @@ function hidingBar.drag:showOnHoverWithDelay()
 	if hidingBar:IsShown() or hidingBar.config.showDelay == 0 then
 		hidingBar:enter()
 	else
-		self:hoverHandler()
-		self.timer = hidingBar.config.showDelay
-		self:SetScript("OnUpdate", self.showBarDelay)
+		hidingBar.drag:hoverHandler()
+		self.fTimer.timer = hidingBar.config.showDelay
+		self.fTimer:SetScript("OnUpdate", self.showBarDelay)
 	end
 end
 
@@ -1379,8 +1533,8 @@ end
 
 
 hidingBar.drag:SetScript("OnLeave", function(self)
-	self:SetScript("OnUpdate", nil)
-	if hidingBar.config.fade and not hidingBar:IsShown() then
+	self.fTimer:SetScript("OnUpdate", nil)
+	if hidingBar.config.fade and not hidingBar:IsShown() and hidingBar.drag:IsShown() then
 		UIFrameFadeOut(self, hidingBar.config.showDelay, self:GetAlpha(), hidingBar.config.fadeOpacity)
 	end
 	hidingBar:leave()